Move useStepText to its own file.
