+++ protocol-designer/src/pages/Designer/ProtocolSteps/Timeline/ConnectedStepInfo.tsx
@@ -1,5 +1,4 @@
 import { useEffect, useMemo } from 'react'
-import { useTranslation } from 'react-i18next'
 import { useDispatch, useSelector } from 'react-redux'
 import debounce from 'lodash/debounce'
 
@@ -10,13 +9,7 @@ import {
   CLOSE_UNSAVED_STEP_FORM,
   ConfirmDeleteModal,
 } from '/protocol-designer/components/organisms'
-import {
-  PAUSE_UNTIL_RESUME,
-  PAUSE_UNTIL_TEMP,
-  PAUSE_UNTIL_TIME,
-} from '/protocol-designer/constants'
 import { selectors as dismissSelectors } from '/protocol-designer/dismiss'
-import { getEnableConcurrentModuleActions } from '/protocol-designer/feature-flags/selectors'
 import { selectors as fileDataSelectors } from '/protocol-designer/file-data'
 import { stepIconsByType } from '/protocol-designer/form-types'
 import { selectors as stepFormSelectors } from '/protocol-designer/step-forms'
@@ -36,8 +29,8 @@ import {
   toggleViewSubstep,
 } from '/protocol-designer/ui/steps/actions/actions'
 
-import { formatTime } from '../../utils'
 import { ConnectedStepContainer } from './ConnectedStepContainer'
+import { useStepText } from './useStepText'
 import {
   getMetaSelectedSteps,
   getMouseClickKeyInfo,
@@ -48,7 +41,7 @@ import {
 import type { ThunkDispatch } from 'redux-thunk'
 import type { Dispatch, MouseEvent, SetStateAction } from 'react'
 import type { DeleteModalType } from '/protocol-designer/components/organisms'
-import type { FormData, StepIdType } from '/protocol-designer/form-types'
+import type { StepIdType } from '/protocol-designer/form-types'
 import type { BaseState, ThunkAction } from '/protocol-designer/types'
 import type { SelectMultipleStepsAction } from '/protocol-designer/ui/steps'
 
@@ -261,35 +254,4 @@ export function ConnectedStepInfo(props: ConnectedStepInfoProps): JSX.Element {
       />
     </>
   )
-}
-
-function useStepText(step: FormData): { text: string; subtext: string | null } {
-  const { i18n, t } = useTranslation(['application', 'protocol_steps'])
-  const enableConcurrentModuleActions = useSelector(
-    getEnableConcurrentModuleActions
-  )
-
-  // add empty check to avoid causing undefined issue when calling titleCase
-  const text =
-    step.stepName !== undefined && step.stepName !== ''
-      ? i18n.format(step.stepName, 'titleCase')
-      : t(`stepType.${step.stepType}`)
-
-  let subtext = null
-  if (enableConcurrentModuleActions && step.stepType === 'pause') {
-    // todo(mm, 2025-09-10): Improve FormData typing to make this type-safe.
-    if (step.pauseAction === PAUSE_UNTIL_RESUME) {
-      subtext = t('protocol_steps:pause.untilResume')
-    } else if (step.pauseAction === PAUSE_UNTIL_TEMP) {
-      subtext = t('protocol_steps:pause.untilTemperature', {
-        temperature: step.pauseTemperature,
-      })
-    } else if (step.pauseAction === PAUSE_UNTIL_TIME) {
-      subtext = t('protocol_steps:pause.forDuration', {
-        duration: formatTime(step.pauseTime as string),
-      })
-    }
-  }
-
-  return { text, subtext }
 }
